Add HSV-colored plotting of flight paths in MATLAB
Also made an attempt at plotting isochrones, but it turns out that the
time series data generated in the simulation isn't as perfect as we
might want. The time intervals aren't consistent (probably due to the
async nature of AirSim). So I'll have to do a little more work to figure
out how to get the isochrones.